#5514 update command args, fix null token

+++ src/main/java/edu/harvard/iq/dataverse/engine/command/impl/MergeInAccountCommand.java
@@ -147,7 +147,9 @@ public class MergeInAccountCommand extends AbstractVoidCommand {
         //  AuthenticatedUserLookup
         //  apiToken
         ApiToken toRemove = ctxt.authentication().findApiTokenByUser(consumedAU);
-        ctxt.em().remove(toRemove);
+        if(null != toRemove) { //not all users have apiTokens
+            ctxt.em().remove(toRemove);
+        }
         AuthenticatedUserLookup consumedAUL = consumedAU.getAuthenticatedUserLookup();
         ctxt.em().remove(consumedAUL);
         ctxt.em().remove(consumedAU);
